commenting; gcc warnings fixed
git-svn-id: file:///home/rhornig/tmp/svn/inet-framework/trunk@873 539eec35-cc16-0410-b457-356aab9780e1

+++ Network/OSPFv2/MessageHandler/LinkStateRequestHandler.cc
@@ -20,13 +20,13 @@ void OSPF::LinkStateRequestHandler::ProcessPacket (OSPFPacket* packet, OSPF::Int
     {
         OSPFLinkStateRequestPacket* lsRequestPacket = check_and_cast<OSPFLinkStateRequestPacket*> (packet);
 
-        unsigned int          requestCount = lsRequestPacket->getRequestsArraySize ();
+        unsigned long         requestCount = lsRequestPacket->getRequestsArraySize ();
         bool                  error        = false;
         std::vector<OSPFLSA*> lsas;
 
         ev << "  Processing packet contents:\n";
 
-        for (int i = 0; i < requestCount; i++) {
+        for (unsigned long i = 0; i < requestCount; i++) {
             LSARequest&      request = lsRequestPacket->getRequests (i);
             OSPF::LSAKeyType lsaKey;
             char             addressString[16];
